Move interactor to namespace
