--- app/interactions/domain_delete_interaction/base.rb
@@ -1,7 +1,0 @@
-module DomainDeleteInteraction
-  class Base < ActiveInteraction::Base
-    object :domain,
-           class: Domain,
-           description: 'Domain to delete'
-  end
-end
