--- app/interactions/domain_delete_interaction/delete.rb
@@ -1,11 +1,0 @@
-module DomainDeleteInteraction
-  class Delete < Base
-    def execute
-      ::PaperTrail.request.whodunnit = "interaction - #{self.class.name}"
-      WhoisRecord.where(domain_id: domain.id).destroy_all
-
-      domain.destroy
-      compose(DomainDeleteInteraction::NotifyRegistrar, inputs)
-    end
-  end
-end
