+++ app/interactions/domains/delete/base.rb
@@ -1,0 +1,9 @@
+module Domains
+  module Delete
+    class Base < ActiveInteraction::Base
+      object :domain,
+             class: Domain,
+             description: 'Domain to delete'
+    end
+  end
+end
