+++ app/jobs/domain_delete_job.rb
@@ -3,6 +3,6 @@ class DomainDeleteJob < Que::Job
   def run(domain_id)
     domain = Domain.find(domain_id)
 
-    DomainDeleteInteraction::Delete.run(domain: domain)
+    Domains::Delete::DoDelete.run(domain: domain)
   end
 end
